pyproject.toml
[project]
name = "a2a-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "a2a-sdk[http-server]>=0.3.25",
    "mcp>=1.26.0",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
]
